+++ tests/Shiny.Device.Tests/Shiny.Device.Tests/BluetoothLE/CharacteristicTests.cs
@@ -6,7 +6,6 @@ using System.Reactive.Linq;
 using System.Reactive.Threading.Tasks;
 using System.Threading.Tasks;
 using Shiny.BluetoothLE;
-using Shiny.BluetoothLE.Central;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -16,7 +15,7 @@ namespace Shiny.Devices.Tests.BluetoothLE
     public class CharacteristicTests : IDisposable
     {
         readonly ITestOutputHelper output;
-        readonly ICentralManager manager;
+        readonly IBleManager manager;
         IGattCharacteristic[] characteristics;
         IPeripheral peripheral;
 
@@ -24,7 +23,7 @@ namespace Shiny.Devices.Tests.BluetoothLE
         public CharacteristicTests(ITestOutputHelper output)
         {
             this.output = output;
-            this.manager = ShinyHost.Resolve<ICentralManager>();
+            this.manager = ShinyHost.Resolve<IBleManager>();
         }