+++ tests/Shiny.Device.Tests/Shiny.Device.Tests/Settings/MyTestEnum.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+
+namespace Shiny.Device.Tests.Settings
+{
+    public enum MyTestEnum
+    {
+        DontKnow,
+        Hi,
+        Bye
+    }
+}